Layout: Move methods to tabify plugins from the main window to it
